Renamed metrics modules
